requirements.txt
pyserial~=3.5
PyYAML~=6.0
matplotlib==3.7.1
pandas~=1.3.4
openpyxl~=3.1.2
email-to
tk
zmq